+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -2,6 +2,7 @@ package nextstep.jdbc;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import javax.sql.DataSource;
 import org.slf4j.Logger;
@@ -16,6 +17,8 @@ public abstract class JdbcTemplate {
     public abstract DataSource getDataSource();
 
     public abstract void setValues(PreparedStatement pstmt) throws SQLException;
+
+    public abstract Object mapRow(ResultSet rs) throws SQLException;
 
     public void update() {
         final DataSource dataSource = getDataSource();
@@ -32,5 +35,28 @@ public abstract class JdbcTemplate {
             log.error(e.getMessage(), e);
             throw new RuntimeException(e);
         }
+    }
+
+    public Object query() {
+        String query = createQuery();
+        DataSource dataSource = getDataSource();
+
+        try (Connection conn = dataSource.getConnection();
+                PreparedStatement pstmt = conn.prepareStatement(query)){
+
+            log.debug("query : {}", query);
+
+            setValues(pstmt);
+            ResultSet rs = executeQuery(pstmt);
+
+            return mapRow(rs);
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        }
+    }
+
+    private ResultSet executeQuery(PreparedStatement pstmt) throws SQLException {
+        return pstmt.executeQuery();
     }
 }
